Moved test folder to src folder
