+++ src/__test__/App.test.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-// import renderer from 'react-test-renderer';
-import App from '../App';
+import App from '../components/App';
 
 test('renders the App with no issue', () => {
   const div = document.createElement('div');
