+++ src/__test__/Button.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import Button from '../Button';
+import Button from '../components/Button';
 
 test('should render the button component', () => {
   render(<Button name="4" wide={true} color='green' handleClick={fireEvent.click} />);
