+++ src/__test__/Calculator.test.js
@@ -1,6 +1,6 @@
 import renderer from 'react-test-renderer';
 import React from 'react';
-import Calculator from '../Calculator';
+import Calculator from '../components/Calculator';
 
 it('matches the Calculator snapshot', () => {
   const tree = renderer.create(<Calculator />).toJSON();
