+++ src/__test__/Display.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import Display from '../Display';
+import Display from '../components/Display';
 
 test('Display the default output of 0', () => {
   const { getByTestId } = render(<Display />);
