+++ src/__test__/Home.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import renderer from 'react-test-renderer';
-import Home from '../Home';
+import Home from '../components/Home';
 
 test('should have a greeting text', () => {
   const { getByTestId } = render(<Home />);
