+++ src/__test__/Navbar.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Navbar from '../Navbar';
+import Navbar from '../components/Navbar';
 
 test('should conatin the organization name', () => {
   const { getByTestId } = render(<Router><Navbar /></Router>);
