+++ src/__test__/Quote.test.js
@@ -1,6 +1,6 @@
 import renderer from 'react-test-renderer';
 import React from 'react';
-import Quote from '../Quote';
+import Quote from '../components/Quote';
 
 it('matches the Quote snapshot', () => {
   const tree = renderer.create(<Quote />).toJSON();
